requirements.txt
pyinstaller
PyQt5
PyQt5-Qt5
PyQt5-sip
pyserial